# backend/swagger.yaml
...
openapi: 3.0.0
info:
  title: BrewBuzz API
  version: 1.0.0
  description: BrewBuzz API endpoints testing.
servers:
  - url: http://localhost:3001

#--Auth User--
paths:
...
  /api/homepage:
    get:
      summary: Get homepage data for the logged-in user
      description: Returns personalized data depending on user category (Customer or Business)
      tags:
        - Home
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Homepage data fetched successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true  # You can replace this with exact structure if needed
        "401":
          description: Unauthorized (invalid or missing token)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "500":
          description: Server error
...
  /api/favorites:
    get:
      summary: Get all favorite businesses for the logged-in user
      tags:
        - Favorites
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of favorite businesses
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  favorites:
                    type: array
                    items:
                      $ref: '#/components/schemas/FavoriteBusiness'
        "500":
          description: Failed to fetch favorites
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FavoriteBusiness:
      type: object
      properties:
        id:
          type: integer
        cover_photo:
          type: string
        name:
          type: string
        location:
          type: string
        phone_number:
          type: string
        description:
          type: string
        averageRating:
          type: number
          format: float
        reviewCount:
          type: integer
...